requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
openai==1.3.0
moviepy==1.0.3
python-dotenv==1.0.0
pillow==10.1.0
loguru==0.7.2 